Add <idea-version/> in plugin.xml with min 203.5981.155
I believe this tag is required for IDEA to allow installing this plugin on any version from specified min and onwards. Without it IDEA seems to restrict min and max to current minor version of the IDE used to build the jar.

Could not build to test sadly, got some complaints about PSI classes being missing.